+++ index.js
@@ -1,7 +1,19 @@
 const Synth = require('./hydra-synth.js')
-const ShaderGenerator = require('./shader-generator.js')
+//const ShaderGenerator = require('./shader-generator.js')
 
-module.exports = {
-  synth : Synth,
-  shaderGenerator: ShaderGenerator
-}
+module.exports = Synth
+
+// console.log('module', module)
+//
+// console.log('exports', exports)
+//
+// module.exports = {
+//     synth : Synth,
+//     shaderGenerator: ShaderGenerator
+//   }
+// // }else{
+// //     var Hydra = {
+// //       synth: Synth,
+// //       shaderGenerator: shaderGenerator
+// //     }
+// // }