+++ CHANGELOG.md
@@ -1,5 +1,16 @@
+2.0.0
+
 - resize() function
-- ShaderGenerator
 - updates to MakeGlobal
 - eval() function on hydra synth
 - dynamically add functions
+- need to require hydra.synth rather than just base repo
+- width and height globally defined
+- update function
+
+
+to do:
+- set framerate
+- ShaderGenerator [WIP], runs serverside
+- check for errors on eval
+- add outputs (o0, etc) and sources to shader generator